STYLE: fix warnings about signed/unsigned mismatches
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=22285
git-svn-id: http://svn.slicer.org/Slicer4/trunk@22285 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/DisplayableManager/vtkMRMLVolumeGlyphSliceDisplayableManager.cxx
@@ -374,7 +374,7 @@ void vtkMRMLVolumeGlyphSliceDisplayableManager::vtkInternal
     }
   // Make sure there is no more display nodes than the volume has
   assert( displayNodesIt->second.size() <=
-          displayNodesIt->first->GetNumberOfDisplayNodes());
+          (unsigned int) displayNodesIt->first->GetNumberOfDisplayNodes());
 }
 
 //---------------------------------------------------------------------------
